+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Currently in development.
 - Improved search query builder
 - Improved HTTP API for users and locations
 - Added support for default values to object reference schemas
+- Added configurable external link dropdowns for objects, actions, instruments, topics, locations, basic groups and project groups
 - Improved handling of new, non-required properties in automatic schema update
 - Added config value SHOW_DOWNTIME_WARNING for warning about occasional downtime on front page
 - Give GRANT permissions to the user creating an object even if also granting permissions to a group